move logscribe types to a new file
