+++ node/pkg/logscribe/api/controller.go
@@ -1,19 +1,9 @@
 package api
 
 import (
-	"encoding/json"
-
 	"github.com/gofiber/fiber/v2"
 	"github.com/rs/zerolog/log"
 )
-
-type LogInsertModel struct {
-	Service   string          `db:"service" json:"service"`
-	Timestamp string          `db:"timestamp" json:"timestamp"`
-	Level     int             `db:"level" json:"level"`
-	Message   string          `db:"message" json:"message"`
-	Fields    json.RawMessage `db:"fields" json:"fields"`
-}
 
 func insertLogs(c *fiber.Ctx) error {
 	defer func(c *fiber.Ctx) {
